src/models/background_gen/gan_bg.py
""" GAN-based background generation implementation """
import torch
import torch.nn as nn
import numpy as np
from PIL import Image
from typing import Dict, Any, Optional
import logging

class SimpleGAN(nn.Module):
        # """Simple GAN for background generation"""
    def __init__(self, latent_dim: int = 100, image_size: int = 512):
        super().__init__()
        self.latent_dim = latent_dim
        self.image_size = image_size

        self.generator = nn.Sequential(
            nn.Linear(latent_dim, 128),
            nn.ReLU(True),
            nn.Linear(128, 256),
            nn.ReLU(True),
            nn.Linear(256, 512),
            nn.ReLU(True),
            nn.Linear(512, image_size * image_size * 3),
            nn.Tanh()
        )

    def forward(self, z):
        return self.generator(z).view(-1, 3, self.image_size, self.image_size)

class GANBackgroundGenerator:
    """Background generator using GAN"""
    def __init__(self, config: Dict[str, Any]):
        self.logger = logging.getLogger(__name__)
        self.config = config
        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")

        # Initialize GAN
        self.latent_dim = config.get('latent_dim', 100)
        self.image_size = min(config.get('resolution', (512, 512)))  # Use smaller dimension

        self.generator = SimpleGAN(self.latent_dim, self.image_size).to(self.device)

        # Load pre-trained weights if available
        model_path = config.get('model_path')
        if model_path:
            try:
                self.generator.load_state_dict(torch.load(model_path, map_location=self.device))
                self.logger.info(f"Loaded GAN model from {model_path}")
            except Exception as e:
                self.logger.warning(f"Failed to load GAN model: {e}")

        self.generator.eval()

    def generate(self, style: str = "default") -> Image.Image:
        """
        Generate background image

        Args:
            style: Style hint for generation

        Returns:
            Generated background image
        """
        try:
            with torch.no_grad():
                # Generate random latent vector
                z = torch.randn(1, self.latent_dim).to(self.device)

                # Generate image
                fake_image = self.generator(z)

                # Convert to PIL Image
                image = fake_image.squeeze(0).cpu()
                image = (image + 1) / 2  # Denormalize from [-1, 1] to [0, 1]
                image = image.clamp(0, 1)

                # Convert to numpy and then PIL
                image_np = image.permute(1, 2, 0).numpy()
                image_np = (image_np * 255).astype(np.uint8)

                pil_image = Image.fromarray(image_np)

                # Resize to target resolution
                target_resolution = self.config.get('resolution', (512, 512))
                pil_image = pil_image.resize(target_resolution, Image.Resampling.LANCZOS)

                return pil_image

        except Exception as e:
            self.logger.error(f"Error generating GAN background: {e}")
            # Return fallback
            return Image.new('RGB', self.config.get('resolution', (512, 512)),
                           color=(135, 206, 235))